laze: init tar support
