+++ go.mod
@@ -3,7 +3,16 @@ module github.com/emcfarlane/laze
 go 1.16
 
 require (
+	github.com/containerd/stargz-snapshotter/estargz v0.7.0
+	github.com/docker/cli v20.10.7+incompatible // indirect
+	github.com/docker/docker v20.10.7+incompatible // indirect
+	github.com/docker/docker-credential-helpers v0.6.4 // indirect
 	github.com/emcfarlane/starlarkassert v0.0.0-20210612114505-0b5ce3fc3821
+	github.com/golang/snappy v0.0.4 // indirect
+	github.com/google/go-containerregistry v0.5.1
+	github.com/klauspost/compress v1.13.1 // indirect
+	github.com/sirupsen/logrus v1.8.1 // indirect
 	go.starlark.net v0.0.0-20210602144842-1cdb82c9e17a
-	golang.org/x/sys v0.0.0-20210616094352-59db8d763f22 // indirect
+	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c // indirect
+	golang.org/x/sys v0.0.0-20210630005230-0f9fa26af87c // indirect
 )
