+++ rule_test.go
@@ -1,0 +1,3 @@
+package laze
+
+// TODO: test rules.
