+++ rules/protoc.star
